instagram-fastapi-backend-main/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
requests
pydantic
openai
gspread
oauth2client
PyPDF2
python-docx

